backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
ultralytics==8.2.13
opencv-python==4.9.0.80
numpy==2.1.2
requests==2.32.3